scripts/pyannote-sidecar/requirements.txt
pyannote.audio==3.1.1
torch>=2.0,<3
torchaudio>=2.0,<3
